chore: create a utils file for download functions
