--- src/node-lib/curriculum-api-2023/queries/lessonDownloads/getNextLessonsInUnit.ts
@@ -1,22 +1,0 @@
-import { LessonListSchema } from "./lessonDownloads.schema";
-
-const getNextLessonsInUnit = (unit: LessonListSchema, lessonSlug: string) => {
-  const lessonInUnit = unit.find((lesson) => lesson.lessonSlug === lessonSlug);
-  const lessonPosition = lessonInUnit?.orderInUnit;
-  const nextLessons =
-    lessonPosition &&
-    unit
-      .filter((lesson) =>
-        lesson.orderInUnit ? lesson.orderInUnit > lessonPosition : [],
-      )
-      .splice(0, 3);
-
-  return nextLessons
-    ? nextLessons.map(({ lessonSlug, lessonTitle }) => ({
-        lessonSlug,
-        lessonTitle,
-      }))
-    : [];
-};
-
-export default getNextLessonsInUnit;
