+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.gql
@@ -3,7 +3,7 @@ query lessonDownloads(
   $programmeSlug: String!
   $unitSlug: String!
 ) {
-  download_assets: published_mv_lesson_content_1_1(
+  download_assets: published_mv_lesson_content_1_3(
     where: { lesson_slug: { _eq: $lessonSlug } }
   ) {
     has_slide_deck_asset_object
@@ -14,6 +14,7 @@ query lessonDownloads(
     starter_quiz
     exit_quiz
     is_legacy
+    expired: deprecated_fields(path: "expired")
   }
 
   unit_lessons: published_mv_synthetic_unitvariant_lessons_by_keystage_6_0_0(
