+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.query.ts
@@ -7,116 +7,16 @@ import errorReporter from "../../../../common-lib/error-reporter";
 import OakError from "../../../../errors/OakError";
 import { Sdk } from "../../sdk";
 
-import getNextLessonsInUnit from "./getNextLessonsInUnit";
 import lessonDownloadsSchema, {
-  LessonDownloadsListSchema,
-  LessonDownloadsPageData,
   downloadsAssetData,
 } from "./lessonDownloads.schema";
+import {
+  constructDownloadsArray,
+  constructLessonListingObjectArray,
+  getNextLessonsInUnit,
+} from "./downloadUtils";
 
-export const getDownloadsArray = (content: {
-  hasSlideDeckAssetObject: boolean;
-  hasStarterQuiz: boolean;
-  hasExitQuiz: boolean;
-  hasWorksheetAssetObject: boolean;
-  hasWorksheetAnswersAssetObject: boolean;
-  hasWorksheetGoogleDriveDownloadableVersion: boolean;
-  hasSupplementaryAssetObject: boolean;
-  isLegacy: boolean;
-}): LessonDownloadsPageData["downloads"] => {
-  const downloads: LessonDownloadsListSchema = [
-    {
-      exists: content.hasSlideDeckAssetObject,
-      type: "presentation",
-      ext: "pptx",
-      label: "Slide deck",
-    },
-    {
-      exists: content.hasStarterQuiz,
-      type: "intro-quiz-questions",
-      label: "Starter quiz questions",
-      ext: "pdf",
-    },
-    {
-      exists: content.hasStarterQuiz,
-      type: "intro-quiz-answers",
-      label: "Starter quiz answers",
-      ext: "pdf",
-    },
-    {
-      exists: content.hasExitQuiz,
-      type: "exit-quiz-questions",
-      label: "Exit quiz questions",
-      ext: "pdf",
-    },
-    {
-      exists: content.hasExitQuiz,
-      type: "exit-quiz-answers",
-      label: "Exit quiz answers",
-      ext: "pdf",
-    },
-    {
-      exists:
-        (!content.isLegacy &&
-          (content.hasWorksheetAssetObject ||
-            content.hasWorksheetAnswersAssetObject)) ||
-        (content.isLegacy &&
-          content.hasWorksheetGoogleDriveDownloadableVersion),
-      type: "worksheet-pdf",
-      label: "Worksheet",
-      ext: "pdf",
-    },
-    {
-      exists:
-        (!content.isLegacy &&
-          (content.hasWorksheetAssetObject ||
-            content.hasWorksheetAnswersAssetObject)) ||
-        (content.isLegacy &&
-          content.hasWorksheetGoogleDriveDownloadableVersion),
-      type: "worksheet-pptx",
-      label: "Worksheet",
-      ext: "pptx",
-    },
-    {
-      exists: content.hasSupplementaryAssetObject,
-      type: "supplementary-pdf",
-      label: "Additional material",
-      ext: "pdf",
-    },
-    {
-      exists: content.hasSupplementaryAssetObject,
-      type: "supplementary-docx",
-      label: "Additional material",
-      ext: "docx",
-    },
-  ];
-
-  return downloads;
-};
-
-const constructLessonListingObjectArray = (
-  unitLessons: SyntheticUnitvariantLessons[],
-) => {
-  const unitLessonsArray = unitLessons.map((lesson) => {
-    return {
-      lessonSlug: lesson.lesson_slug,
-      lessonTitle: lesson.lesson_data.title,
-      description: lesson.lesson_data.description,
-      pupilLessonOutcome: lesson.lesson_data.pupil_lesson_outcome,
-      expired: lesson.lesson_data.deprecated_fields ? true : false,
-      quizCount:
-        (lesson.lesson_data.quiz_id_starter ? 1 : 0) +
-        (lesson.lesson_data.quiz_id_exit ? 1 : 0),
-      videoCount: lesson.lesson_data.video_id ? 1 : 0,
-      presentationCount: lesson.lesson_data.asset_id_slidedeck ? 1 : 0,
-      worksheetCount: lesson.lesson_data.asset_id_worksheet ? 1 : 0,
-      hasCopyrightMaterial: lesson.lesson_data.copyright_content ? true : false,
-      orderInUnit: lesson.supplementary_data.order_in_unit,
-      lessonCohort: lesson.lesson_data._cohort,
-    };
-  });
-  return unitLessonsArray;
-};
+// ADD TEST FOR CORRECT RESOURCES
 
 const lessonDownloadsQuery =
   (sdk: Sdk) =>
@@ -126,6 +26,7 @@ const lessonDownloadsQuery =
     lessonSlug: string;
   }) => {
     const res = await sdk.lessonDownloads(args);
+
     if (
       !res.download_assets ||
       !res.unit_lessons ||
@@ -133,6 +34,17 @@ const lessonDownloadsQuery =
       res.unit_lessons.length === 0
     ) {
       throw new OakError({ code: "curriculum-api/not-found" });
+    }
+
+    if (res.download_assets.length > 1) {
+      const error = new OakError({
+        code: "curriculum-api/uniqueness-assumption-violated",
+      });
+      errorReporter("curriculum-api-2023::lessonDownloads")(error, {
+        severity: "warning",
+        ...args,
+        res,
+      });
     }
 
     const { download_assets, unit_lessons } = res;
@@ -146,22 +58,12 @@ const lessonDownloadsQuery =
       starter_quiz,
       exit_quiz,
       is_legacy,
+      expired,
     } = downloadsAssetData.parse(download_assets[0]);
 
-    const { programme_slug } = syntheticUnitvariantLessonsSchema.parse(
+    const parsedUnitLessons = syntheticUnitvariantLessonsSchema.parse(
       unit_lessons[0],
     );
-
-    const { updated_at } = syntheticUnitvariantLessonsSchema.parse(
-      unit_lessons[0],
-    ).lesson_data;
-
-    const { subject, subject_slug, keystage_slug, keystage_description } =
-      syntheticUnitvariantLessonsSchema.parse(unit_lessons[0]).programme_fields;
-
-    const { title, slug } = syntheticUnitvariantLessonsSchema.parse(
-      unit_lessons[0],
-    ).unit_data;
 
     const downloadsData = {
       hasSlideDeckAssetObject: has_slide_deck_asset_object,
@@ -175,39 +77,28 @@ const lessonDownloadsQuery =
       isLegacy: is_legacy,
     };
 
-    const downloads = getDownloadsArray(downloadsData);
+    const downloads = constructDownloadsArray(downloadsData);
 
-    const pageData = {
-      downloads,
-      programmeSlug: programme_slug,
-      keyStageSlug: keystage_slug,
-      keyStageTitle: keystage_description,
-      lessonSlug: args.lessonSlug,
-      lessonTitle: title,
-      subjectSlug: subject_slug,
-      subjectTitle: subject,
-      unitSlug: slug,
-      unitTitle: title,
-      lessonCohort: null,
-      expired: null,
-      updatedAt: updated_at,
-    };
-
-    if (res.download_assets.length > 1) {
-      const error = new OakError({
-        code: "curriculum-api/uniqueness-assumption-violated",
-      });
-      errorReporter("curriculum-api-2023::lessonDownloads")(error, {
-        severity: "warning",
-        ...args,
-        res,
-      });
-    }
     const unitLessonsArray = constructLessonListingObjectArray(
       res.unit_lessons as SyntheticUnitvariantLessons[],
     );
-
     const nextLessons = getNextLessonsInUnit(unitLessonsArray, args.lessonSlug);
+
+    const pageData = {
+      downloads,
+      programmeSlug: parsedUnitLessons.programme_slug,
+      keyStageSlug: parsedUnitLessons.programme_fields.keystage_slug,
+      keyStageTitle: parsedUnitLessons.programme_fields.keystage_description,
+      lessonSlug: parsedUnitLessons.lesson_slug,
+      lessonTitle: parsedUnitLessons.lesson_data.title,
+      subjectSlug: parsedUnitLessons.programme_fields.subject_slug,
+      subjectTitle: parsedUnitLessons.programme_fields.subject,
+      unitSlug: parsedUnitLessons.unit_data.slug,
+      unitTitle: parsedUnitLessons.unit_data.title,
+      lessonCohort: parsedUnitLessons.lesson_data._cohort,
+      expired: expired ? expired : null,
+      updatedAt: parsedUnitLessons.lesson_data.updated_at,
+    };
 
     return lessonDownloadsSchema.parse({
       ...pageData,
